HEAD - findbugs changes

+++ org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/JavaModelCache.java
@@ -155,7 +155,7 @@ public IJavaElement getExistingElement(IJavaElement element) {
 }
 
 protected double getMemoryRatio() {
-	if (this.memoryRatio == -1) {
+	if ((int) this.memoryRatio == -1) {
 		long maxMemory = Runtime.getRuntime().maxMemory();
 		// if max memory is infinite, set the ratio to 4d which corresponds to the 256MB that Eclipse defaults to
 		// (see https://bugs.eclipse.org/bugs/show_bug.cgi?id=111299)
